+++ TP/tp2/question2/FahrenheitCelsius.java
@@ -54,7 +54,7 @@ public class FahrenheitCelsius{
             
 	    
           
-	    float c1= (((float)5)/((float)9));
+	    float c1= (float)5/9;
 	  
 	    float conv = c1 *  (entree - 32);
